depends/docker-registry-core/tests/requirements.txt
nose==1.3.3
coverage==3.7.1
mock==1.0.1
mockredispy==2.9.0.6
